2021 02 05 글쓰기화면에 글쓴이 JSTL로 가져오기 테스트완료/ 주석처리완료 /freeboard/register.jsp에 session 얻어와서 user라는이름으로 memberVO객체를 SET해줌

+++ src/main/java/org/zerock/domain/MemberVO.java
@@ -4,6 +4,6 @@ import lombok.Data;
 
 @Data
 public class MemberVO {
-	private Long no;
+	private int no;
 	private String name;
 }
